smooth operator was not working to changed the output, will check again tomorrow, task 3.1 code written for single alpha for loop not plotting yet
Smooth operator was not working i got help to correct it now it works. Task 3.1 for loop not working yet but individual values of alpha show smoothening curve. Need to compare the two codes in smooth operator. your previous code is there as a comment. Will do this tomorrow. Old Task 3.1 when you used k is now in removed code file.